chore: remove dead code and unused deps flagged by an over-engineering audit
Deletes unused helper functions across sankey.py, ctutils.py, config.py,
verbose_utils.py, result_cache.py, simulation_worker.py, and cantera_docs.py
that had zero callers anywhere in the repo. Factors a duplicated YAML
extension check in config.py into one helper. Drops the unused
@tanstack/react-query and class-variance-authority frontend dependencies
(and the now-pointless QueryClientProvider wrapper), removes an identity-map
lookup table (KIND_LABELS) in favor of using the kind string directly, folds
a duplicated API-error-parsing helper into the shared apiFetch wrapper, and
inlines the single-consumer useKeyboardShortcuts hook into AppShell. Also
moves the mypy-only types-PyYAML stub package to the dev dependency group
instead of runtime dependencies.

make qa, make type-check, targeted/full backend pytest, frontend
type-check/unit tests, and a production frontend build all pass.

Extensive AI use — code & PR fully written by Claude Sonnet 5.
