+++ boulder/sankey.py
@@ -499,149 +499,6 @@ def generate_sankey_input_from_sim(
     # https://github.com/oemof/tespy/blob/dd0059c0d993c00d8d99fc87de1e4246ec6a684d/src/tespy/tools/analyses.py#L810
 
 
-# Functions to edit Sankey diagrams ;
-# for instance emulate recirculations
-
-
-def get_outlet_value(links, nodes, node_name, filter_links="", get_color=False):
-    """Get the sum of all outlet streams of a node.
-
-    Also returns the color of the largest link.
-
-    Parameters
-    ----------
-    links : dict
-        Network links. Dictionary with keys {'source', 'target', 'value', 'color', 'label'}
-    nodes : list
-        Network list of names of nodes.
-    node_name : str
-        Name of the node.
-    filter_links : str
-        Expression to capture name of output streams to aggregate. Default is "",
-        i.e. all output streams are aggregated. Example::
-
-            filter_links = "H2|CH4"
-    get_color: str
-        If True, get color of the largest link. The default is False.
-    """
-    idx = nodes.index(node_name)
-    # all links with source = idx
-    all_links = [i for i, s in enumerate(links["source"]) if s == idx]
-    # filter with regex filter_links:
-    if filter_links:
-        import re
-
-        all_links = [i for i in all_links if re.match(filter_links, links["label"][i])]
-
-    # sort by value
-    all_links.sort(key=lambda i: links["value"][i], reverse=True)
-
-    # sum values for all outlet streams:
-    value = sum([links["value"][i] for i in all_links])
-
-    if get_color:
-        color = links["color"][all_links[0]]
-        return value, color
-    else:
-        return value
-
-
-def add_link(links, nodes, source_str, target_str, value, color=None, label=None):
-    """Add a connection between Source and Target.
-
-    Parameters
-    ----------
-    links : dict
-        Dictionary of links, with keys {'source', 'target', 'value', 'color', 'label'}
-    nodes : list
-        list of names of nodes.
-    source_str : str
-        Name of the source node.
-    target_str : str
-        Name of the target node.
-    value : float
-        Value of the link.
-    color : str, optional
-        Color of the link. The default is None.
-    label : str, optional
-        Label of the link. The default is None.
-    """
-    # Add the link
-    assert source_str in nodes, f"source_str not in nodes: {source_str}"
-    assert target_str in nodes, f"target_str not in nodes: {target_str}"
-    links["source"].append(nodes.index(source_str))
-    links["target"].append(nodes.index(target_str))
-    links["value"].append(value)
-    if color is None:
-        color = "grey"
-    links["color"].append(color)
-    if label is None:
-        label = ""
-    links["label"].append(label)
-
-
-def substract_value(
-    links, nodes, source_str, target_str, value, link_name=None, allow_negative=False
-):
-    """Substract a value from a link.
-
-    Parameters
-    ----------
-    links : dict
-        Dictionary of links, with keys {'source', 'target', 'value', 'color', 'label'}
-    nodes : list
-        list of names of nodes.
-    source_str : str
-        Name of the source node.
-    target_str : str
-        Name of the target node.
-    value : float
-        Value to substract from existing link.
-    link_name : str, optional
-        Name of the link. The default is None.
-    allow_negative : bool, optional
-        Allow negative values. The default is False. If True, the value being subtracted
-        cannot be greater than the existing value.
-
-    Returns
-    -------
-    None
-        links and nodes are modified in place.
-    """
-    # Find the index of the link
-    if link_name is None:
-        idx = [
-            i
-            for i, (s, t) in enumerate(zip(links["source"], links["target"]))
-            if s == nodes.index(source_str) and t == nodes.index(target_str)
-        ]
-        assert len(idx) == 1, (
-            f"Found {len(idx)} links between {source_str} and {target_str}"
-        )
-        idx = idx[0]
-    else:  # find link by "link_name"
-        # assert link name exists
-        assert link_name in links["label"], f"Link name not found: {link_name}"
-        idx = [
-            i
-            for i, (s, t) in enumerate(zip(links["source"], links["target"]))
-            if s == nodes.index(source_str) and t == nodes.index(target_str)
-        ]
-        # filter by link name
-        idx = [i for i in idx if links["label"][i] == link_name]
-        assert len(idx) == 1, (
-            f"Found {len(idx)} links between {source_str} and {target_str} with label {link_name}"
-        )
-        idx = idx[0]
-
-    # Substract the value
-    if not allow_negative:
-        assert links["value"][idx] >= value, (
-            f"Value to substract is greater than existing value: {value} > {links['value'][idx]}"
-        )
-    links["value"][idx] -= value
-
-
 if __name__ == "__main__":
     # Smoke test: use a plugin-provided default simulation if available.
     # The plugin package must register a module path via entry point
