+++ frontend/src/App.tsx
@@ -1,21 +1,11 @@
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { Toaster } from "sonner";
 import { AppShell } from "@/components/layout/AppShell";
 
-const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      retry: 1,
-      staleTime: 30_000,
-    },
-  },
-});
-
 export default function App() {
   return (
-    <QueryClientProvider client={queryClient}>
+    <>
       <AppShell />
       <Toaster position="bottom-right" richColors />
-    </QueryClientProvider>
+    </>
   );
 }
