+++ frontend/src/components/layout/AppShell.test.tsx
@@ -30,10 +30,6 @@ vi.mock("@/hooks/useSimulationSSE", () => ({
   useSimulationSSE: vi.fn(),
 }));
 
-vi.mock("@/hooks/useKeyboardShortcuts", () => ({
-  useKeyboardShortcuts: vi.fn(),
-}));
-
 vi.mock("@/api/configs", () => ({
   fetchPreloadedConfig: vi.fn().mockResolvedValue({ preloaded: false }),
   fetchDefaultConfig: vi.fn().mockResolvedValue({ config: { nodes: [], connections: [] }, yaml: "" }),
