+++ pyproject.toml
@@ -17,7 +17,6 @@ dependencies = [
   "h5py>=3.0.0",
   "python-dotenv>=1.0.0",
   "PyYAML>=6.0",
-  "types-PyYAML>=6.0.0",
   "pydantic>=2.0.0",
   "fastapi>=0.110.0",
   "uvicorn[standard]>=0.27.0",
@@ -34,7 +33,8 @@ readme = "README.md"
 dev = [
   "mypy>=1.0.0",
   "httpx>=0.27.0",
-  "pytest-asyncio>=0.23.0"
+  "pytest-asyncio>=0.23.0",
+  "types-PyYAML>=6.0.0"
 ]
 
 [project.scripts]
